backend/src/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
nest-asyncio>=1.5.8
IPython>=8.12.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
asyncio>=3.4.3
python-multipart>=0.0.6
requests>=2.31.0
pytz
python-docx>=0.8.11
pandas>=2.2.0
python-socketio>=5.8.0
